BravePipeExtractor: RUMBLE: skip 'auto' stream as it provides a master HLS playlist
We don't support it - as in NewPipe there is no automatic bandwidth
changing implemented AFAIK

If selected it still would work and I guess the best quality will be played
in the player. But in the download section it is redundant as we already have
all the variants there to choose from.